p1_navigation/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.9.0=pypi_0
argon2-cffi=20.1.0=py36he774522_1
astor=0.8.1=pypi_0
atomicwrites=1.4.0=pypi_0
attrs=19.3.0=py_0
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=1.5.0=pypi_0
ca-certificates=2020.6.24=0
certifi=2020.6.20=py36_0
cffi=1.14.0=py36h7a1dbc1_0
colorama=0.4.3=py_0
cycler=0.10.0=py36h009560c_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
docopt=0.6.2=pypi_0
entrypoints=0.3=py36_0
freetype=2.10.2=hd328e21_0
gast=0.4.0=pypi_0
grpcio=1.11.0=pypi_0
html5lib=0.9999999=pypi_0
icu=58.2=ha925a31_3
importlib-metadata=1.7.0=py36_0
importlib_metadata=1.7.0=0
iniconfig=1.0.1=pypi_0
intel-openmp=2020.1=216
ipykernel=5.3.4=py36h5ca1d4c_0
ipython=7.16.1=py36h5ca1d4c_0
ipython-genutils=0.2.0=pypi_0
ipython_genutils=0.2.0=py36h3c5d0ee_0
ipywidgets=7.5.1=py_0
jedi=0.17.2=py36_0
jinja2=2.11.2=py_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py36_0
jupyter=1.0.0=py36_7
jupyter_client=6.1.6=py_0
jupyter_console=6.1.0=py_0
jupyter_core=4.6.3=py36_0
kiwisolver=1.2.0=py36h74a9793_0
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.18=h62dcd97_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.2.2=pypi_0
markupsafe=1.1.1=py36he774522_0
matplotlib=3.2.2=0
matplotlib-base=3.2.2=py36h64f37c6_0
mistune=0.8.4=py36he774522_0
mkl=2020.1=216
mkl-service=2.3.0=py36hb782905_0
mkl_fft=1.1.0=py36h45dec08_0
mkl_random=1.1.1=py36h47e9c7a_0
more-itertools=8.4.0=pypi_0
msys2-conda-epoch=20160418=1
nbconvert=5.6.1=py36_0
nbformat=5.0.7=py_0
notebook=6.1.1=py36_0
numpy=1.19.1=py36h5510c5b_0
numpy-base=1.19.1=py36ha3acd2a_0
openssl=1.1.1g=he774522_1
packaging=20.4=py_0
pandas=1.1.0=py36ha925a31_0
pandoc=2.10.1=0
pandocfilters=1.4.2=py36_1
parso=0.7.1=pypi_0
pickleshare=0.7.5=pypi_0
pillow=7.2.0=pypi_0
pip=20.2.2=py36_0
pluggy=0.13.1=pypi_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.6=pypi_0
prompt_toolkit=3.0.5=0
protobuf=3.5.2=pypi_0
py=1.9.0=pypi_0
pycparser=2.20=py_2
pygments=2.6.1=py_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py36h6538335_2
pyrsistent=0.16.0=py36he774522_0
pytest=6.0.1=pypi_0
python=3.6.10=h9f7ef89_2
python-dateutil=2.8.1=py_0
pytorch=0.4.0=py36_cuda80_cudnn7he774522_1
pytz=2020.1=py_0
pywin32=227=py36he774522_1
pywinpty=0.5.7=py36_0
pyyaml=5.3.1=pypi_0
pyzmq=19.0.1=py36ha925a31_1
qt=5.9.7=vc14h73c81de_0
qtconsole=4.7.5=py_0
qtpy=1.9.0=py_0
send2trash=1.5.0=py36_0
setuptools=49.6.0=py36_0
sip=4.19.8=py36h6538335_0
six=1.15.0=py_0
sqlite=3.32.3=h2a8f88b_0
tensorboard=1.7.0=pypi_0
tensorflow=1.7.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.8.3=py36_0
testpath=0.4.4=py_0
toml=0.10.1=pypi_0
tornado=6.0.4=py36he774522_1
tqdm=4.48.2=py_0
traitlets=4.3.3=py36_0
unityagents=0.4.0=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wcwidth=0.2.5=py_0
webencodings=0.5.1=py36_1
werkzeug=1.0.1=pypi_0
wheel=0.34.2=py36_0
widgetsnbextension=3.5.1=py36_0
wincertstore=0.2=py36h7fe50ca_0
winpty=0.4.3=4
zeromq=4.3.2=ha925a31_2
zipp=3.1.0=py_0
zlib=1.2.11=h62dcd97_4

pytest~=6.0.1
numpy~=1.19.1
tensorflow~=1.7.1
unityagents~=0.4.0
pyyaml~=5.3.1
docopt~=0.6.2
setuptools~=49.6.0
pillow~=7.2.0
matplotlib~=3.2.2